refactor(node): share backend helpers
